BuildScript/requirements.txt
PyInstaller>=5.0.0
Pillow>=9.0.0 